+++ code/crates/core-consensus/src/handle/proposed_value.rs
@@ -3,8 +3,6 @@ use crate::prelude::*;
 use crate::handle::driver::apply_driver_input;
 use crate::types::{ProposedValue, WalEntry};
 
-use super::decide::try_decide;
-
 /// Handles a proposed value that can originate from multiple sources:
 /// 1. Application layer:
 ///    - In 'parts-only' mode
@@ -85,12 +83,6 @@ where
             DriverInput::Proposal(signed_proposal, validity),
         )
         .await?;
-    }
-
-    if origin == ValueOrigin::Sync {
-        // The proposed value was provided by Sync, try to decide immediately, without waiting for the Commit timeout.
-        // `try_decide` will check that we are in the commit step after applying the proposed value to the state machine.
-        try_decide(co, state, metrics).await?;
     }
 
     Ok(())