+++ code/crates/starknet/test/src/lib.rs
@@ -172,7 +172,7 @@ impl TestRunner {
                     ..Default::default()
                 },
                 max_tx_count: 10000,
-                gossip_batch_size: 100,
+                gossip_batch_size: 0,
                 load: MempoolLoadConfig {
                     load_type: MempoolLoadType::UniformLoad(UniformLoadConfig::default()),
                 },
@@ -189,7 +189,10 @@ impl TestRunner {
                     .unwrap(),
             },
             runtime: RuntimeConfig::single_threaded(),
-            test: TestConfig::default(),
+            test: TestConfig {
+                stable_block_times: true,
+                ..TestConfig::default()
+            },
         }
     }
 }
@@ -232,7 +235,6 @@ fn apply_params(config: &mut Config, params: &TestParams) {
     config.consensus.p2p.protocol = params.protocol;
     config.consensus.timeouts.timeout_step = params.timeout_step;
     config.test.max_block_size = params.block_size;
-    config.test.tx_size = params.tx_size;
     config.test.txs_per_part = params.txs_per_part;
     config.test.vote_extensions.enabled = params.vote_extensions.is_some();
     config.test.vote_extensions.size = params.vote_extensions.unwrap_or_default();